\name{add_interactivity.points}
\alias{add_interactivity.points}
\title{Add interactivity to points grob}
\usage{
  add_interactivity.points(gr, data)
}
\arguments{
  \item{gr}{A points grob}

  \item{data}{The data containing interactivity values}
}
\value{
  A garnished points grob
}
\description{
  Add interactivity to points grob
}
\author{
  chris
}

