\name{add_javascript}
\alias{add_javascript}
\title{Add javascript files to the plot}
\usage{
  add_javascript(grob)
}
\arguments{
  \item{grob}{A garnished Grob}
}
\value{
  Garnished Grob with additional scriptGrob children
}
\description{
  add_javascript This function adds scriptGrobs to a given
  grob. The javascript files to be added are those which
  are necessary for adding interactivity to a plot
}

