\name{as_svgmap.SpatialPixelsDataFrame}
\alias{as_svgmap.SpatialPixelsDataFrame}
\title{as_svgmaps.SpatialPixelsDataFrame}
\usage{
  as_svgmap.SpatialPixelsDataFrame(object, ...)
}
\arguments{
  \item{sp_pixels_df}{A SpatialPixelsDataFrame object}

  \item{vars}{Variables to be kept}
}
\value{
  Data frame in svgmaps format
}
\description{
  Transforms SpatialPixelsDataFrame into svgmaps data frame
}
\author{
  chris
}

