\name{iopts}
\alias{iopts}
\title{iopts}
\usage{
  iopts(...)
}
\arguments{
  \item{...}{The named options}
}
\value{
  An ioptions object
}
\description{
  Set interactive properties
}
\details{
  Set interactive options. Like the options in ggplot2,
  this can be used to control the plot appearance. This
  controls the interactive porperties. Currently only
  hcolour is implemented. Hcolour defines the colour in
  which the elements are highlighted.
}
\author{
  chris
}

