\name{melt_nodes}
\alias{melt_nodes}
\title{Melt nodes}
\usage{
  melt_nodes(osmar_obj)
}
\arguments{
  \item{osmar_obj}{an osmar object}
}
\value{
  dataframe containing coordinates and attributes
}
\description{
  Merges the attributes and tags of nodes with a subset of
  the key variable.
}
\details{
  content
}
\author{
  chris
}

