\docType{data}
\name{muc}
\alias{muc}
\title{Munich's central city}
\format{An osmar object}
\description{
  An excerpt of munichs central city. The data originally
  comes from OpenStreetMap and is also in the datasets of
  the osmar package.
}
\references{
  \url{http://cran.r-project.org/web/packages/osmar/index.html}
}
\keyword{datasets}

