\name{save_svgmap}
\alias{save_svgmap}
\title{Save svgmap}
\usage{
  save_svgmap(object, filename = "RPlot.svg")
}
\arguments{
  \item{object}{svgmap object}

  \item{filename}{the filename of the SVG file}
}
\description{
  Saves a svgmap object as SVG.
}
\details{
  This function adds the necessary scripts for
  interactivity and converts the plot into an SVG file.
}
\author{
  chris
}

