\docType{data}
\name{signals}
\alias{signals}
\title{Traffic signal points}
\format{A svgmap data frame}
\description{
  A svgmap data frame containing only nodes which represent
  the traffic signals in a bounding box around munich
  (germany)
}
\references{
  \url{openstreetmap.de/}
}
\keyword{datasets}

