\docType{data}
\name{signals_grid}
\alias{signals_grid}
\title{Traffic signal tiles}
\format{A SpatialPixelsDataFrame}
\description{
  A SpatialPixelsDataFrame containing the traffic signals
  counts on a regular grid overlayed over munich (germany)
}
\references{
  \url{openstreetmap.de/}
}
\keyword{datasets}

