\name{+.svgmap}
\alias{+.svgmap}
\title{"+.svgmap"}
\usage{
  \method{+}{svgmap} (p, object)
}
\arguments{
  \item{p}{plot object}

  \item{object}{A object which shall be added to the plot}
}
\value{
  A svgmap object
}
\description{
  Modify a plot by adding new layers
}
\details{
  Same as in ggplot2 but for svgmap()
}
\author{
  chris
}

