\name{as_svgmap.SpatialGridDataFrame}
\alias{as_svgmap.SpatialGridDataFrame}
\title{as_svgmaps.SpatialGridDataFrame}
\usage{
  \method{as_svgmap}{SpatialGridDataFrame} (object, ...)
}
\arguments{
  \item{object}{A SpatialGridDataFrame object}

  \item{...}{Not in use}
}
\value{
  Data Frame in svgmaps format
}
\description{
  Transforms SpatialGridDataFrame into svgmaps data frame
}
\details{
  This function internally uses the
  as_svgmaps.SpatialPixelsDataFrame function
}
\author{
  chris
}

