\name{as_svgmap.SpatialLinesDataFrame}
\alias{as_svgmap.SpatialLinesDataFrame}
\title{as_svgmaps.SpatialLinesDatFrame}
\usage{
  as_svgmap.SpatialLinesDataFrame(object, ...)
}
\arguments{
  \item{sp_lines_df}{A SpatialLinesDataFrame object}

  \item{vars}{Variables to be kept}
}
\value{
  Data frame in svgmaps format
}
\description{
  Transforms SpatialLinesDataFrame into svgmaps data frame
}
\details{
  This function merges the coordinates and the data of a
  SpatialLinesDataFrame
}
\author{
  chris
}

