\name{as_svgmap.SpatialPixelsDataFrame}
\alias{as_svgmap.SpatialPixelsDataFrame}
\title{as_svgmaps.SpatialPixelsDataFrame}
\usage{
  \method{as_svgmap}{SpatialPixelsDataFrame} (object, ...)
}
\arguments{
  \item{object}{A SpatialPixelsDataFrame object}

  \item{...}{Not in use}
}
\value{
  Data frame in svgmaps format
}
\description{
  Transforms SpatialPixelsDataFrame into svgmaps data frame
}
\author{
  chris
}

